Rearrange Python API (utils)
Split up utils.py into a separate "utils" module.
