--- docs/api_reference/utils.md
@@ -1,3 +1,0 @@
-# utils
-
-::: ci_cd.utils
