+++ docs/api_reference/utils/console_printing.md
@@ -1,0 +1,3 @@
+# console_printing
+
+::: ci_cd.utils.console_printing
